refactor: Add getSpacingStyle utility function
