+++ demo/vite.config.ts
@@ -1,15 +1,15 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import path from "path";
-import { vanillaExtractPlugin } from "@vanilla-extract/vite-plugin";
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react(), vanillaExtractPlugin()],
+  plugins: [react()],
   resolve: {
     alias: {
       "@layer-ui": path.resolve(__dirname, "../packages/ui"),
       "@layer-lib": path.resolve(__dirname, "../packages/lib"),
+      "@layer-utils": path.resolve(__dirname, "../packages/utils"),
     },
   },
   build: {
