--- packages/lib/use-spacing/src/index.ts
@@ -1,2 +1,0 @@
-"use client";
-export { useSpacing } from "./use-spacing";
