+++ tsconfig.json
@@ -30,7 +30,6 @@
       "@layer-lib/react-context": ["./packages/lib/react-context/src"],
       "@layer-lib/react-use-composed-ref": ["./packages/lib/use-composed-ref/src"],
       "@layer-lib/react-use-browser-layout-effect": ["./packages/lib/use-browser-layout-effect/src"],
-      "@layer-lib/use-spacing": ["./packages/lib/use-spacing/src"],
       "@layer-lib/react-use-escape-key-down": ["./packages/lib/use-escape-key-down/src"],
 
       // ui
@@ -38,6 +37,9 @@
       "@layer-ui/portal": ["./packages/ui/portal/src"],
       "@layer-ui/modal": ["./packages/ui/modal/src"],
       "@layer-ui/presence": ["./packages/ui/presence/src"],
+
+      // utils
+      "@layer-utils/getSpacingStyle": ["./packages/utils/getSpacingStyle/src"],
     }
   },
   "include": ["demo", "packages", "ssr"]
